+++ backend/model/player.py
@@ -1,4 +1,4 @@
-from model.hand import Hand
+from backend.model.hand import Hand
 from pydantic import BaseModel, Field
 from typing import List
 from collections import Counter
