+++ backend/tests/test_agent.py
@@ -1,6 +1,6 @@
 import pytest
-from model.game import Game
-from model.player import Player
+from backend.model.game import Game
+from backend.model.player import Player
 from backend.game_controller import GameController
 from backend.agent.agent import (
     predict,
