+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
 
 [tool.pytest.ini_options]
 pythonpath = ["./"]
-asyncio_default_fixture_loop_scope = "function"
+asyncio_default_fixture_loop_scope = "gfunction"
 
 [tool.ruff.lint]
 ignore = ["F405", "F403"]
